fix: ntp response check stratum value
NTP uses the concept of a stratum to describe how many NTP hops away
a machine is from an authoritative time source.

For example, a stratum 1 time server has a radio or atomic clock
directly attached to it. It then sends its time to a stratum 2 time
server through NTP, and so on.

A machine running NTP automatically chooses the machine with the lowest
stratum number that it is configured to communicate with using NTP
as its time source.

This strategy effectively builds a self-organizing tree of NTP
speakers. NTP performs well over the non-deterministic path lengths of
packet-switched networks, because it makes robust estimates of the
following three key variables in the relationship between a client
and a time server.

stratum between 1 and 15 is acceptable

http://www.cisco.com/c/en/us/support/docs/availability/high-availability/19643-ntpm.html#highstratum