fix: remove maxConcurrent in flatmap
This doesn't exactly work in a way you would expect to favorably
expect.  If you have only 3 slots, then the error-ing hosts are retried
immediately blocking any future ones from following through.

we could have a circular buffer, removing the errored out ones and
putting them back in a pool but given how tedious that will be, don't
think it's worth making that change. if anyone requests a PR for it, i
can probably try pulling it off with a Subject source